Started Adding helpers
